Split out config loading and parsing.
Having a separate class for finding and loading config files will make
separating out a standalone library later much easier.
