+++ src/AssetConfig.php
@@ -46,13 +46,6 @@ class AssetConfig
     );
 
     /**
-     * The timestamp that configuration changed.
-     *
-     * @var integer
-     */
-    protected $_modifiedTime;
-
-    /**
      * A hash of constants that can be expanded when reading ini files.
      *
      * @var array
@@ -82,7 +75,7 @@ class AssetConfig
      */
     public function __construct(array $data = array(), array $additionalConstants = array(), $modifiedTime = null)
     {
-        $this->_data = $data;
+        $this->_data = $data ?: static::$_defaults;
         $this->constantMap = array_merge($this->constantMap, $additionalConstants);
         if (!$modifiedTime) {
             $modifiedTime = time();
@@ -91,19 +84,61 @@ class AssetConfig
     }
 
     /**
-     * Constructor
+     * Factory method
      *
      * @param string $iniFile File path for the ini file to parse.
      * @param array $additionalConstants  Additional constants that will be translated
      *    when parsing paths.
+     * @deprecated Use ConfigFinder::loadAll() instead.
      */
     public static function buildFromIniFile($iniFile = null, $constants = array())
     {
         if (empty($iniFile)) {
             $iniFile = CONFIG . 'asset_compress.ini';
         }
+        $config = new static([], $constants);
+        return $config->load($iniFile);
+    }
 
-        return self::_parseConfig($iniFile, $constants);
+    /**
+     * Load a config file into the current instance.
+     *
+     * @param string $path The config file to load.
+     * @param string $prefix The string to prefix all targets in $path with.
+     * @return $this
+     */
+    public function load($path, $prefix = '')
+    {
+        $config = $this->_readConfig($path);
+
+        foreach ($config as $section => $values) {
+            if (in_array($section, self::$_extensionTypes)) {
+                // extension section, merge in the defaults.
+                $defaults = $this->get($section);
+                if ($defaults) {
+                    $values = array_merge($defaults, $values);
+                }
+                $this->addExtension($section, $values);
+
+            } elseif (strtolower($section) === self::GENERAL) {
+                $this->set(self::GENERAL, $values);
+
+            } elseif (strpos($section, self::FILTER_PREFIX) === 0) {
+                // filter section.
+                $name = str_replace(self::FILTER_PREFIX, '', $section);
+                $this->filterConfig($name, $values);
+
+            } else {
+                $lastDot = strrpos($section, '.') + 1;
+                $extension = substr($section, $lastDot);
+                $key = $section;
+
+                // must be a build target.
+                $this->addTarget($prefix . $key, $values);
+            }
+        }
+
+        return $this;
     }
 
     /**
@@ -113,7 +148,7 @@ class AssetConfig
      * @return array Inifile contents
      * @throws RuntimeException
      */
-    protected static function _readConfig($filename)
+    protected function _readConfig($filename)
     {
         if (empty($filename) || !is_string($filename) || !file_exists($filename)) {
             throw new RuntimeException(sprintf('Configuration file "%s" was not found.', $filename));
@@ -123,96 +158,6 @@ class AssetConfig
             return parse_ini_file($filename, true);
         } else {
             return parse_ini_string(file_get_contents($filename), true);
-        }
-    }
-
-    /**
-     * Transforms the config data into a more structured form
-     *
-     * @param array $contents Contents to build a config object from.
-     * @param array $constants Array of constants that will be mapped.
-     * @param int $modifiedTime The modified time of the config data.
-     * @return AssetConfig
-     */
-    protected static function _parseConfig($baseFile, $constants, $modifiedTime = null)
-    {
-        if (!$modifiedTime && file_exists($baseFile)) {
-            $modifiedTime = filemtime($baseFile);
-        }
-
-        $assetConfig = new AssetConfig(self::$_defaults, $constants, $modifiedTime);
-        self::_parseConfigFileLocal($baseFile, $assetConfig);
-
-        $plugins = Plugin::loaded();
-        foreach ($plugins as $plugin) {
-            $pluginConfig = Plugin::path($plugin) . 'config' . DS . 'asset_compress.ini';
-            if (file_exists($pluginConfig)) {
-                self::_parseConfigFileLocal($pluginConfig, $assetConfig, $plugin . '.');
-            }
-        }
-
-        if ($assetConfig->general('cacheConfig')) {
-            self::_writeCache(self::CACHE_ASSET_CONFIG_KEY, $assetConfig);
-        }
-
-        return $assetConfig;
-    }
-
-    /**
-     * Parse a file and optionally the .local version of the file.
-     *
-     * @param string $file The file to parse.
-     * @param AssetConfig $assetConfig The config object to add data to.
-     * @param null|string $prefix The prefix to append to build targets.
-     * @return void
-     */
-    protected static function _parseConfigFileLocal($file, $assetConfig, $prefix = null)
-    {
-        self::_parseConfigFile($file, $assetConfig, $prefix);
-
-        // Load related .local.ini file if exists
-        $localConfig = preg_replace('/(.*)\.ini$/', '$1.local.ini', $file);
-        if (file_exists($localConfig)) {
-            self::_parseConfigFile($localConfig, $assetConfig, $prefix);
-        }
-    }
-
-    /**
-     * Reads a config file and applies it to the given config instance
-     *
-     * @param string $iniFile Contents to apply to the config instance.
-     * @param AssetConfig $assetConfig The config instance instance we're applying this config file to.
-     * @param string $prefix Prefix for the target key
-     */
-    protected static function _parseConfigFile($iniFile, $assetConfig, $prefix = '')
-    {
-        $config = self::_readConfig($iniFile);
-
-        foreach ($config as $section => $values) {
-            if (in_array($section, self::$_extensionTypes)) {
-                // extension section, merge in the defaults.
-                $defaults = $assetConfig->get($section);
-                if ($defaults) {
-                    $values = array_merge($defaults, $values);
-                }
-                $assetConfig->addExtension($section, $values);
-
-            } elseif (strtolower($section) === self::GENERAL) {
-                $assetConfig->set(self::GENERAL, $values);
-
-            } elseif (strpos($section, self::FILTER_PREFIX) === 0) {
-                // filter section.
-                $name = str_replace(self::FILTER_PREFIX, '', $section);
-                $assetConfig->filterConfig($name, $values);
-
-            } else {
-                $lastDot = strrpos($section, '.') + 1;
-                $extension = substr($section, $lastDot);
-                $key = $section;
-
-                // must be a build target.
-                $assetConfig->addTarget($prefix . $key, $values);
-            }
         }
     }
 
@@ -585,15 +530,5 @@ class AssetConfig
     {
         $ext = $this->getExt($target);
         return !empty($this->_data[$ext][self::TARGETS][$target]);
-    }
-
-    /**
-     * Get the modified time of the config object.
-     *
-     * @return integer
-     */
-    public function modifiedTime()
-    {
-        return $this->_modifiedTime;
     }
 }
