+++ tests/TestCase/AssetConfigTest.php
@@ -23,53 +23,6 @@ class AssetConfigTest extends TestCase
         $this->config = AssetConfig::buildFromIniFile($this->testConfig);
     }
 
-    /**
-     * Test local configuration files.
-     *
-     * @return void
-     */
-    public function testLocalConfig()
-    {
-        $ini = dirname($this->testConfig) . DS . 'overridable.ini';
-        $config = AssetConfig::buildFromIniFile($ini);
-
-        $this->assertEquals('', $config->general('cacheConfig'));
-        $this->assertEquals(1, $config->general('alwaysUseController'));
-
-        $this->assertEquals('', $config->get('js.timestamp'));
-
-        $result = $config->paths('js');
-        $result = str_replace('/', DS, $result);
-        $expectedJsPaths = array(WWW_ROOT . 'js' . DS . '*', WWW_ROOT . 'js_local' . DS . '*');
-        $this->assertEquals($expectedJsPaths, $result);
-
-        $this->assertEquals(WWW_ROOT . 'cache_js/', $config->cachePath('js'));
-
-        $result = $config->filters('js');
-        $expectedJsFilters = array('sprockets', 'jsyuicompressor', 'mylocalfilter');
-        $this->assertEquals($expectedJsFilters, $result);
-
-        $result = $config->filterConfig('jsyuicompressor');
-        $this->assertEquals(array('path' => '/path/to/local/yuicompressor'), $result);
-
-        $result = $config->filterConfig('uglify');
-        $this->assertEquals(array('path' => '/path/to/uglify-js'), $result);
-
-        $result = $config->paths('js', 'libs.js');
-        $result = str_replace('/', DS, $result);
-        $expectedJsPaths[] = WWW_ROOT . 'js' . DS . 'libs' . DS . '*';
-        $this->assertEquals($expectedJsPaths, $result);
-
-        $result = $config->files('libs.js');
-        $expected = array('jquery.js', 'mootools.js', 'class.js');
-        $this->assertEquals($expected, $result);
-
-        $result = $config->filters('js', 'libs.js');
-        $expectedJsFilters[] = 'uglify';
-        $expectedJsFilters[] = 'anotherlocalfilter';
-        $this->assertEquals($expectedJsFilters, $result);
-    }
-
     public function testBuildFromIniFile()
     {
         $config = AssetConfig::buildFromIniFile($this->testConfig);
@@ -300,11 +253,5 @@ class AssetConfigTest extends TestCase
     {
         $this->assertTrue($this->config->exists('libs.js'));
         $this->assertFalse($this->config->exists('derped.js'));
-    }
-
-    public function testModifiedTime()
-    {
-        $this->assertInternalType('integer', $this->config->modifiedTime());
-        $this->assertEquals(filemtime($this->testConfig), $this->config->modifiedTime());
     }
 }
